+++ crates/conductor_lib/src/lib.rs
@@ -99,6 +99,7 @@ extern crate log;
 extern crate serde_derive;
 #[macro_use]
 extern crate serde_json;
+#[cfg(test)]
 #[macro_use]
 extern crate maplit;
 #[macro_use]