+++ release/publish/default.nix
@@ -71,7 +71,6 @@ for crate in \
  conductor_api \
  dpki \
  sim2h \
- sim1h \
  net \
  core \
  conductor_lib \